backend/requirements.txt
django>=4.2
djangorestframework
psycopg2-binary
gunicorn
django-cors-headers
djangorestframework-simplejwt